// src/app/core/models/company-profile.model.ts
export interface CompanyProfile {
  id?: string;
  companyId?: string;
  companyTitle: string;
  ticketSubtitle?: string;
  rnc?: string;
  telefono?: string;
  direccion?: string;
  logoUrl?: string;
  logoStoragePath?: string;
  createdAt?: string;
  createdBy?: string;
  updatedAt?: string;
  updatedBy?: string;
}